+++ NET1814_MilkShop.API/Controllers/OrderController.cs
@@ -36,43 +36,6 @@ namespace NET1814_MilkShop.API.Controllers
             return ResponseExtension.Result(response);
         }
 
-        #region OrderHistory
-
-        [HttpGet]
-        [Route("/api/customer/orders")]
-        [Authorize(AuthenticationSchemes = "Access", Roles = "3")]
-        [ServiceFilter(typeof(UserExistsFilter))]
-        public async Task<IActionResult> GetOrderHistory([FromQuery] OrderHistoryQueryModel model)
-        {
-            _logger.Information("Get order history");
-            var userId = (HttpContext.Items["UserId"] as Guid?)!.Value;
-            var res = await _orderService.GetOrderHistoryAsync(userId, model);
-            return ResponseExtension.Result(res);
-        }
-
-        [HttpGet("/api/customer/orders/{id}")]
-        [Authorize(AuthenticationSchemes = "Access", Roles = "3")]
-        [ServiceFilter(typeof(UserExistsFilter))]
-        public async Task<IActionResult> GetOrderHistoryDetail(Guid id)
-        {
-            _logger.Information("Get order detail history");
-            var userId = (HttpContext.Items["UserId"] as Guid?)!.Value;
-            var res = await _orderService.GetOrderHistoryDetailAsync(userId, id);
-            return ResponseExtension.Result(res);
-        }
-
-        [HttpPatch("/api/customer/orders/{id}/cancel")]
-        [Authorize(AuthenticationSchemes = "Access", Roles = "3")]
-        [ServiceFilter(typeof(UserExistsFilter))]
-        public async Task<IActionResult> CancelOrder(Guid id)
-        {
-            _logger.Information("Cancel order");
-            var userId = (HttpContext.Items["UserId"] as Guid?)!.Value;
-            var res = await _orderService.CancelOrderAsync(userId, id);
-            return ResponseExtension.Result(res);
-        }
-
-        #endregion
         [HttpPatch]
         [Route("/api/dashboard/orders/{id}/status")]
         [Authorize(AuthenticationSchemes = "Access", Roles = "1,2")]
