+++ NET1814_MilkShop.API/Controllers/UserController.cs
@@ -4,6 +4,7 @@ using NET1814_MilkShop.API.CoreHelpers.ActionFilters;
 using NET1814_MilkShop.API.CoreHelpers.Extensions;
 using NET1814_MilkShop.Repositories.CoreHelpers.Constants;
 using NET1814_MilkShop.Repositories.Models.AddressModels;
+using NET1814_MilkShop.Repositories.Models.OrderModels;
 using NET1814_MilkShop.Repositories.Models.UserModels;
 using NET1814_MilkShop.Services.Services;
 using ILogger = Serilog.ILogger;
@@ -20,6 +21,7 @@ namespace NET1814_MilkShop.API.Controllers
         private readonly IUserService _userService;
         private readonly ICustomerService _customerService;
         private readonly IAddressService _addressService;
+        private readonly IOrderService _orderService;
 
         public UserController(ILogger logger, IServiceProvider serviceProvider)
         {
@@ -27,6 +29,7 @@ namespace NET1814_MilkShop.API.Controllers
             _userService = serviceProvider.GetRequiredService<IUserService>();
             _customerService = serviceProvider.GetRequiredService<ICustomerService>();
             _addressService = serviceProvider.GetRequiredService<IAddressService>();
+            _orderService = serviceProvider.GetRequiredService<IOrderService>();
         }
 
         #region User
@@ -252,6 +255,44 @@ namespace NET1814_MilkShop.API.Controllers
             return Ok(response);*/
             return ResponseExtension.Result(response);
         }
+        #endregion
+        
+        #region OrderHistory
+
+        [HttpGet]
+        [Route("/api/customer/orders")]
+        [Authorize(AuthenticationSchemes = "Access", Roles = "3")]
+        [ServiceFilter(typeof(UserExistsFilter))]
+        public async Task<IActionResult> GetOrderHistory([FromQuery] OrderHistoryQueryModel model)
+        {
+            _logger.Information("Get order history");
+            var userId = (HttpContext.Items["UserId"] as Guid?)!.Value;
+            var res = await _orderService.GetOrderHistoryAsync(userId, model);
+            return ResponseExtension.Result(res);
+        }
+
+        [HttpGet("/api/customer/orders/{id}")]
+        [Authorize(AuthenticationSchemes = "Access", Roles = "3")]
+        [ServiceFilter(typeof(UserExistsFilter))]
+        public async Task<IActionResult> GetOrderHistoryDetail(Guid id)
+        {
+            _logger.Information("Get order detail history");
+            var userId = (HttpContext.Items["UserId"] as Guid?)!.Value;
+            var res = await _orderService.GetOrderHistoryDetailAsync(userId, id);
+            return ResponseExtension.Result(res);
+        }
+
+        [HttpPatch("/api/customer/orders/{id}/cancel")]
+        [Authorize(AuthenticationSchemes = "Access", Roles = "3")]
+        [ServiceFilter(typeof(UserExistsFilter))]
+        public async Task<IActionResult> CancelOrder(Guid id)
+        {
+            _logger.Information("Cancel order");
+            var userId = (HttpContext.Items["UserId"] as Guid?)!.Value;
+            var res = await _orderService.CancelOrderAsync(userId, id);
+            return ResponseExtension.Result(res);
+        }
+
         #endregion
     }
 }
